core: breakup `externs.d.ts`
