+++ cli/cli-flags.js
@@ -351,7 +351,7 @@ function getFlags(manualArgv, options = {}) {
   // Augmenting yargs type with auto-camelCasing breaks in tsc@4.1.2 and @types/yargs@15.0.11,
   // so for now cast to add yarg's camelCase properties to type.
   const argv = /** @type {Awaited<typeof parser.argv>} */ (parser.argv);
-  const cliFlags = /** @type {typeof argv & CamelCasify<typeof argv>} */ (argv);
+  const cliFlags = /** @type {typeof argv & LH.Util.CamelCasify<typeof argv>} */ (argv);
 
   // yargs will return `undefined` for options that have a `coerce` function but
   // are not actually present in the user input. Instead of passing properties
