+++ core/config/budget.js
@@ -135,9 +135,9 @@ class Budget {
    * Returns the budget that applies to a given URL.
    * If multiple budgets match based on thier 'path' property,
    * then the last-listed of those budgets is returned.
-   * @param {Immutable<Array<LH.Budget>>|null} budgets
+   * @param {LH.Util.Immutable<Array<LH.Budget>>|null} budgets
    * @param {string|undefined} url
-   * @return {Immutable<LH.Budget> | undefined} budget
+   * @return {LH.Util.Immutable<LH.Budget> | undefined} budget
    */
   static getMatchingBudget(budgets, url) {
     if (budgets === null || url === undefined) return;
