+++ types/externs.d.ts
@@ -8,101 +8,6 @@ import {Artifacts} from './artifacts';
 import LHResult from './lhr/lhr';
 import {SharedFlagsSettings, OutputMode} from './lhr/settings';
 
-declare global {
-  // Augment Intl to include
-  // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/getCanonicalLocales
-  namespace Intl {
-    var getCanonicalLocales: (locales?: string | Array<string>) => Array<string>;
-  }
-
-  interface Window {
-    // Cached native functions/objects for use in case the page overwrites them.
-    // See: `executionContext.cacheNativesOnNewDocument`.
-    __nativePromise: PromiseConstructor;
-    __nativePerformance: Performance;
-    __nativeFetch: typeof fetch,
-    __nativeURL: typeof URL;
-    __ElementMatches: Element['matches'];
-    __HTMLElementBoundingClientRect: HTMLElement['getBoundingClientRect'];
-
-    /** Used for monitoring long tasks in the test page. */
-    ____lastLongTask?: number;
-
-    /** Used by FullPageScreenshot gatherer. */
-    __lighthouseNodesDontTouchOrAllVarianceGoesAway: Map<Element, string>;
-    __lighthouseExecutionContextUniqueIdentifier?: number;
-
-    /** Injected into the page when the `--debug` flag is used. */
-    continueLighthouseRun(): void;
-
-    // Not defined in tsc yet: https://github.com/microsoft/TypeScript/issues/40807
-    requestIdleCallback(callback: (deadline: {didTimeout: boolean, timeRemaining: () => DOMHighResTimeStamp}) => void, options?: {timeout: number}): number;
-  }
-
-  /** Make properties K in T optional. */
-  type MakeOptional<T, K extends keyof T> = {
-    [P in Exclude<keyof T, K>]: T[P]
-  } & {
-    [P in K]+?: T[P]
-  }
-
-  /** An object with the keys in the union K mapped to themselves as values. */
-  type SelfMap<K extends string> = {
-    [P in K]: P;
-  };
-
-  /** Make optional all properties on T and any properties on object properties of T. */
-  type RecursivePartial<T> =
-    // Recurse into arrays and tuples: elements aren't (newly) optional, but any properties they have are.
-    T extends (infer U)[] ? RecursivePartial<U>[] :
-    // Recurse into objects: properties and any of their properties are optional.
-    T extends object ? {[P in keyof T]?: RecursivePartial<T[P]>} :
-    // Strings, numbers, etc. (terminal types) end here.
-    T;
-
-  /** Recursively makes all properties of T read-only. */
-  type Immutable<T> =
-    T extends Function ? T :
-    T extends Array<infer R> ? ImmutableArray<R> :
-    T extends Map<infer K, infer V> ? ImmutableMap<K, V> :
-    T extends Set<infer M> ? ImmutableSet<M> :
-    T extends object ? ImmutableObject<T> :
-    T
-
-  // Intermediate immutable types. Prefer e.g. Immutable<Set<T>> over direct use.
-  type ImmutableArray<T> = ReadonlyArray<Immutable<T>>;
-  type ImmutableMap<K, V> = ReadonlyMap<Immutable<K>, Immutable<V>>;
-  type ImmutableSet<T> = ReadonlySet<Immutable<T>>;
-  type ImmutableObject<T> = {
-    readonly [K in keyof T]: Immutable<T[K]>;
-  };
-
-  /**
-   * Exclude void from T
-   */
-  type NonVoid<T> = T extends void ? never : T;
-
-  /** Remove properties K from T. */
-  type StrictOmit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
-
-  /** Obtain the type of the first parameter of a function. */
-  type FirstParamType<T extends (arg1: any, ...args: any[]) => any> =
-    T extends (arg1: infer P, ...args: any[]) => any ? P : never;
-
-  /**
-   * If `S` is a kebab-style string `S`, convert to camelCase.
-   */
-  type KebabToCamelCase<S> =
-    S extends `${infer T}-${infer U}` ?
-    `${T}${Capitalize<KebabToCamelCase<U>>}` :
-    S
-
-  /** Returns T with any kebab-style property names rewritten as camelCase. */
-  type CamelCasify<T> = {
-    [K in keyof T as KebabToCamelCase<K>]: T[K];
-  }
-}
-
 /**
  * Extends the flags in SharedFlagsSettings with flags used to configure the
  * Lighthouse module but will not end up in the Config settings.
